% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_generator.R
\name{mater}
\alias{mater}
\title{Create a single recombinant offspring from two randomly sampled parents}
\usage{
mater(mat, loci, ploidy = 1)
}
\arguments{
\item{mat}{a matrix of individuals x alleles}

\item{loci}{a factor deliniating the loci in each column}

\item{ploidy}{the ploidy of the sample}
}
\value{
a vector of alleles
}
\description{
Create a single recombinant offspring from two randomly sampled parents
}
\examples{
mat <- t(rbind(rmultinom(100, 2, runif(10)),
             rmultinom(100, 2, runif(10)))) # two loci with 20 alleles
loc <- factor(rep(LETTERS[1:2], each = 10))
t(vapply(seq_len(nrow(mat)), function(i) mater(mat, loc, 2), mat[1, ]))
}
